Refactor/Update/Fix Script
This commit changes/fixes the following things:

- Download newest TETR.IO version*
- Move desktop file template into install script and replacement inline
- Remove wget from dependencies in favor of curl
- Added some error checks for dependencies and directories that need to
  be present
- Icons are placed into correct directory
- Fix desktop file `Version` value

*Sadly osk decided to not stick with an URL that points to the latest
TETR.IO version. This means that the version now has to be hard coded
unless i want to scape the HTML and extract the version that way. I
don't think it's worth the effort so I'll just have to live with the
fact that this script has to be updated every now and then. Maybe in the
future I'm annoyed enough to fix it.